fix Windows, more refactoring
