+++ rust/bin-shared/Cargo.toml
@@ -30,6 +30,7 @@ libc = "0.2"
 
 [target.'cfg(target_os = "windows")'.dependencies]
 known-folders = "1.1.0"
+ring = "0.17"
 uuid = { version = "1.10.0", features = ["v4"] }
 wintun = "0.4.0"
