+++ rust/bin-shared/benches/tunnel.rs
@@ -41,13 +41,7 @@ mod platform {
 
     pub(crate) async fn perf() -> Result<()> {
         // Install wintun so the test can run
-        let wintun_path = connlib_shared::windows::wintun_dll_path().unwrap();
-        tokio::fs::create_dir_all(wintun_path.parent().unwrap())
-            .await
-            .unwrap();
-        tokio::fs::write(&wintun_path, connlib_shared::windows::wintun_bytes())
-            .await
-            .unwrap();
+        let _wintun_path = firezone_bin_shared::windows::ensure_dll().unwrap();
 
         const MTU: usize = 1_280;
         const NUM_REQUESTS: u64 = 1_000;
