+++ rust/headless-client/Cargo.toml
@@ -54,7 +54,6 @@ dirs = "5.0.1"
 [target.'cfg(target_os = "windows")'.dependencies]
 ipconfig = "0.3.2"
 known-folders = "1.1.0"
-ring = "0.17"
 thiserror = { version = "1.0", default-features = false }
 tracing-subscriber = { version = "0.3.17", features = ["env-filter"] }
 windows-service = "0.7.0"
