+++ rust/headless-client/src/windows.rs
@@ -10,9 +10,6 @@ use std::path::{Path, PathBuf};
 pub(crate) use socket_factory::tcp as tcp_socket_factory;
 pub(crate) use socket_factory::udp as udp_socket_factory;
 
-#[path = "windows/wintun_install.rs"]
-mod wintun_install;
-
 // The return value is useful on Linux
 #[allow(clippy::unnecessary_wraps)]
 pub(crate) fn check_token_permissions(_path: &Path) -> Result<()> {
@@ -34,6 +31,6 @@ pub(crate) fn notify_service_controller() -> Result<()> {
 }
 
 pub(crate) fn setup_before_connlib() -> Result<()> {
-    wintun_install::ensure_dll()?;
+    firezone_bin_shared::windows::ensure_dll()?;
     Ok(())
 }
